requirements.txt
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
openpyxl>=3.0.7
beautifulsoup4==4.12.2
selenium==4.16.0
pdfplumber==0.10.3
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.16.3
pytest>=6.2.5
black>=21.7b0
flake8>=3.9.2
linkedin-scraper==1.0.3
webdriver-manager==4.0.1 